+++ CHANGELOG.md
@@ -9,6 +9,7 @@ customers cannot upgrade their bootloader, its changes are recorded separately.
 ### [Unreleased]
 - Bitcoin: add support for sending to silent payment (BIP-352) addresses
 - Bitcoin: add support for regtest
+- Cardano: add support for vote delegation
 
 ### 9.20.0
 - Bitcoin: UX improvements for payment request confirmations